test: reorganize backend client extensions (WPB-26603)
