+++ tests/testsSupport/src/main/backendUtils/BackendClient.kt
@@ -24,39 +24,12 @@
 package backendUtils
 
 import CredentialsManager
-import android.net.Uri
-import backendUtils.team.TeamRoles
-import backendUtils.team.defaultheaders
-import backendUtils.team.getAuthToken
-import backendUtils.team.getTeamId
 import com.wire.android.testSupport.BuildConfig
-import com.wire.android.testSupport.backendConnections.team.Team
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.runBlocking
-import logger.WireTestLogger
-import network.HttpRequestException
-import network.NetworkBackendClient
-import network.NumberSequence
-import network.RequestOptions
-import org.json.JSONArray
-import org.json.JSONObject
-import service.models.Connection
-import service.models.Conversation
-import service.models.TeamMember
-import user.utils.AccessCookie
-import user.utils.AccessCredentials
-import user.utils.AccessToken
 import user.utils.BasicAuth
-import user.utils.ClientUser
 import java.net.Authenticator
-import java.net.HttpCookie
-import java.net.HttpURLConnection
 import java.net.InetSocketAddress
 import java.net.PasswordAuthentication
 import java.net.Proxy
-import java.net.URI
-import java.net.URL
-import java.net.URLEncoder
 
 @Suppress("LongParameterList", "MagicNumber", "TooManyFunctions")
 class BackendClient(
@@ -198,1000 +171,7 @@ class BackendClient(
         }
     }
 
-    fun createPersonalUserViaBackend(user: ClientUser): ClientUser {
-        WireTestLogger.getLog(NetworkBackendClient::class.simpleName ?: "Null").info("user is $user")
-        val url = URL(this.backendUrl + "register")
-
-        val requestBody = JSONObject().apply {
-            put("email", user.email)
-            put("name", user.name)
-            put("password", user.password)
-        }
-
-        val headers = mapOf(contentType to applicationJson)
-
-        val response = NetworkBackendClient.sendJsonRequestWithCookies(
-            url = url,
-            method = "POST",
-            body = requestBody.toString(),
-            headers = headers
-        )
-
-        val json = JSONObject(response.body)
-        user.id = json.getString("id")
-        val accessCookie = AccessCookie("zuid", response.cookies)
-        user.accessCredentials = AccessCredentials(null, accessCookie)
-
-        val activationCode = getActivationCodeForEmail(
-            user.email.orEmpty()
-        )
-        WireTestLogger.getLog(NetworkBackendClient::class.simpleName ?: "Null")
-            .info("code is $activationCode")
-        activateRegisteredEmailByBackendCode(
-            user.email.orEmpty(),
-            activationCode
-        )
-
-        return user
-    }
-
-    fun createWirelessUserViaBackend(user: ClientUser): ClientUser {
-        val url = URL(this.backendUrl)
-
-        val requestBody = JSONObject().apply {
-            put("name", user.name)
-        }
-
-        if (user.expiresIn != null) {
-            requestBody.put("expires_in", user.expiresIn!!.seconds)
-        }
-
-        val headers = mapOf(contentType to applicationJson)
-
-        val response = NetworkBackendClient.sendJsonRequest(
-            url = url,
-            method = "POST",
-            body = requestBody.toString(),
-            headers = headers
-        )
-
-        val connection = url.openConnection() as HttpURLConnection
-        val cookiesHeader = connection.getHeaderField("Set-Cookie")
-        val cookies = HttpCookie.parse(cookiesHeader).toList()
-
-        val json = JSONObject(response)
-        user.id = json.getString("id")
-        val accessCookie = AccessCookie("zuid", cookies)
-        user.accessCredentials = AccessCredentials(null, accessCookie)
-
-        val activationCode = getActivationCodeForEmail(
-            user.email.orEmpty()
-        )
-        WireTestLogger.getLog(NetworkBackendClient::class.simpleName ?: "Null").info("code is $activationCode")
-        activateRegisteredEmailByBackendCode(user.email.orEmpty(), activationCode)
-        return user
-    }
-
-    fun getActivationCodeForEmail(email: String): String {
-        val encodedEmail = URLEncoder.encode(email, "UTF-8")
-        val url = URL("${backendUrl}i/users/activation-code?email=$encodedEmail")
-        val headers = mapOf(
-            AUTHORIZATION to basicAuth.getEncoded(),
-            "Accept" to applicationJson
-        )
-        val response = NetworkBackendClient.sendJsonRequest(
-            url = url,
-            method = "GET",
-            body = null,
-            headers = headers
-        )
-        return JSONObject(response).getString("code")
-    }
-
-    @Suppress("TooGenericExceptionCaught")
-    fun trigger2FA(email: String) {
-        val url = URL("${backendUrl}v5/verification-code/send")
-
-        val requestBody = JSONObject().apply {
-            put("action", "login")
-            put("email", email)
-        }
-
-        val headers = mapOf(
-            AUTHORIZATION to basicAuth.getEncoded(),
-            applicationJson to applicationJson,
-            "Accept" to applicationJson
-        )
-
-        try {
-            NetworkBackendClient.sendJsonRequest(
-                url = url,
-                method = "POST",
-                body = requestBody.toString(),
-                headers = headers
-            )
-        } catch (e: Exception) {
-            // Optional: Allow HTTP 429 as acceptable like in Java
-        }
-    }
-
-    @Suppress("TooGenericExceptionThrown")
-    fun getVerificationCode(user: ClientUser): String {
-        trigger2FA(user.email.orEmpty())
-
-        val encodedUserId = Uri.encode(user.id)
-        val url = URL("${backendUrl}i/users/$encodedUserId/verification-code/login")
-
-        val headers = mapOf(
-            AUTHORIZATION to basicAuth.getEncoded(),
-            "Accept" to applicationJson
-        )
-
-        return try {
-            val response = NetworkBackendClient.sendJsonRequest(
-                url = url,
-                method = "GET",
-                headers = headers,
-                options = RequestOptions(
-                    expectedResponseCodes = NumberSequence.Range(200..299)
-                )
-            )
-            response.replace("\"", "")
-        } catch (e: Exception) {
-            throw RuntimeException("Failed to get verification code: ${e.message}", e)
-        }
-    }
-
-    // Used to active a register user email
-    fun activateRegisteredEmailByBackendCode(email: String, code: String): String {
-
-        val url = URL("${backendUrl}activate")
-
-        val requestBody = JSONObject().apply {
-            put("email", email)
-            put("code", code)
-            put("dryrun", false)
-        }
-        WireTestLogger.getLog(NetworkBackendClient::class.simpleName ?: "Null").info("JsonBody is $requestBody")
-
-        val headers = mapOf(
-            contentType to applicationJson,
-            "Accept" to applicationJson
-        )
-
-        NetworkBackendClient.sendJsonRequest(
-            url = url,
-            method = "POST",
-            body = requestBody.toString(),
-            headers = headers
-        )
-
-        WireTestLogger.getLog(NetworkBackendClient::class.simpleName ?: "Null")
-            .info("JsonBody response is $requestBody")
-
-        return "Email Registered"
-    }
-
-    fun createIdentityProvider(user: ClientUser, metadata: String): String {
-        val token = runBlocking { getAuthToken(user) }
-        val url = URL("identity-providers".composeCompleteUrl())
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put("Authorization", "${token?.type} ${token?.value}")
-            put("Accept", applicationJson)
-            put("Content-Type", "application/xml")
-        }
-
-        val response = NetworkBackendClient.sendJsonRequestWithCookies(
-            url = url,
-            method = "POST",
-            body = metadata,
-            headers = headers,
-            options = RequestOptions(
-                accessToken = token,
-                expectedResponseCodes = NumberSequence.Array(intArrayOf(HttpURLConnection.HTTP_OK, HttpURLConnection.HTTP_CREATED))
-            )
-        )
-
-        val responseBody = JSONObject(response.body)
-        return responseBody.getString("id")
-    }
-
-    fun getAllTeams(forUser: ClientUser): List<Team> {
-        val token = runBlocking { getAuthToken(forUser) }
-        val url = URL("teams".composeCompleteUrl())
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put("Authorization", "${token?.type} ${token?.value}")
-            put("Accept", applicationJson)
-        }
-
-        val response = NetworkBackendClient.sendJsonRequestWithCookies(
-            url = url,
-            method = "GET",
-            headers = headers,
-            options = RequestOptions(
-                accessToken = token,
-                expectedResponseCodes = NumberSequence.Array(intArrayOf(HttpURLConnection.HTTP_OK))
-            )
-        )
-
-        val jsonResponse = JSONObject(response.body)
-        val teams = jsonResponse.getJSONArray("teams")
-
-        return buildList {
-            for (i in 0 until teams.length()) {
-                add(Team.fromJSON(teams.getJSONObject(i)))
-            }
-        }
-    }
-
-    fun getBackendClientIds(forUser: ClientUser): List<String> {
-        val token = runBlocking { getAuthToken(forUser) }
-        val url = URI("clients".composeCompleteUrl()).toURL()
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put("Authorization", "${token?.type} ${token?.value}")
-            put("Accept", applicationJson)
-        }
-
-        val response = NetworkBackendClient.sendJsonRequestWithCookies(
-            url = url,
-            method = "GET",
-            headers = headers,
-            options = RequestOptions(
-                accessToken = token,
-                expectedResponseCodes = NumberSequence.Array(intArrayOf(HttpURLConnection.HTTP_OK))
-            )
-        )
-
-        val clients = JSONArray(response.body)
-        return buildList {
-            for (i in 0 until clients.length()) {
-                add(clients.getJSONObject(i).getString("id"))
-            }
-        }
-    }
-
-    fun removeBackendClient(forUser: ClientUser, clientId: String) {
-        val token = runBlocking { getAuthToken(forUser) }
-        val url = URI("clients/$clientId".composeCompleteUrl()).toURL()
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put("Authorization", "${token?.type} ${token?.value}")
-            put("Accept", applicationJson)
-        }
-
-        NetworkBackendClient.makeRequest(
-            url = url,
-            method = "DELETE",
-            body = JSONObject().apply {
-                put("password", forUser.password)
-            }.toString(),
-            headers = headers,
-            options = RequestOptions(
-                accessToken = token,
-                expectedResponseCodes = NumberSequence.Array(
-                    intArrayOf(
-                        HttpURLConnection.HTTP_OK,
-                        HttpURLConnection.HTTP_NO_CONTENT
-                    )
-                )
-            )
-        )
-    }
-
-    fun getTeamMembers(asUser: ClientUser): List<TeamMember> {
-        val firstTeam = getAllTeams(asUser).first()
-        return getTeamMembers(runBlocking { getAuthToken(asUser)!! }, firstTeam.id)
-    }
-
-    private fun getTeamMembers(token: AccessToken, teamId: String): List<TeamMember> {
-        val url = URL("teams/$teamId/members".composeCompleteUrl())
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put("Authorization", "${token.type} ${token.value}")
-            put("Accept", applicationJson)
-        }
-
-        val response = NetworkBackendClient.sendJsonRequestWithCookies(
-            url = url,
-            method = "GET",
-            headers = headers,
-            options = RequestOptions(
-                accessToken = token,
-                expectedResponseCodes = NumberSequence.Array(intArrayOf(HttpURLConnection.HTTP_OK))
-            )
-        )
-
-        val jsonResponse = JSONObject(response.body)
-        val members = jsonResponse.getJSONArray("members")
-
-        return buildList {
-            for (i in 0 until members.length()) {
-                val member = members.getJSONObject(i)
-                val userId = member.getString("user")
-                val permissions = member.getJSONObject("permissions")
-                val role = TeamRoles.getByPermissionBitMask(permissions.getInt("self"))
-                add(TeamMember(userId, role))
-            }
-        }
-    }
-
-    private fun getFeatureConfig(feature: String, user: ClientUser): JSONObject {
-        val token = runBlocking {
-            getAuthToken(user)
-        }
-        val url = URL(String.format("feature-configs/%s", feature).composeCompleteUrl())
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put("Authorization", "${token?.type} ${token?.value}")
-        }
-
-        val response = NetworkBackendClient.sendJsonRequestWithCookies(
-            url = url,
-            method = "GET",
-            headers = headers,
-            options = RequestOptions(
-                accessToken = token
-            )
-        )
-
-        val objectResponse = JSONObject(response.body)
-        return objectResponse
-    }
-
-    suspend fun sendConnectionRequest(fromUser: ClientUser, toUser: ClientUser) {
-        val token = getAuthToken(fromUser)
-        val url =
-            URL("connections/${BackendClient.loadBackend(toUser.backendName.orEmpty()).domain}/${toUser.id}".composeCompleteUrl())
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put(AUTHORIZATION, "${token?.type} ${token?.value}")
-        }
-
-        // First try the new endpoint
-        val response = try {
-            NetworkBackendClient.sendJsonRequestWithCookies(
-                url = url,
-                method = "POST",
-                headers = headers,
-                options = RequestOptions(accessToken = token)
-            )
-        } catch (e: Exception) {
-            if (e.message?.contains("404") == true) {
-                // Fallback to old endpoint
-                val fallbackUrl = URL("connections".composeCompleteUrl())
-                val requestBody = JSONObject().apply {
-                    put("user", toUser.id)
-                    put("name", toUser.name)
-                    put("message", "This message is not shown anywhere anymore")
-                }
-
-                NetworkBackendClient.sendJsonRequestWithCookies(
-                    url = fallbackUrl,
-                    method = "POST",
-                    headers = headers,
-                    body = requestBody.toString(),
-                    options = RequestOptions(accessToken = token)
-                )
-            } else {
-                // Retry after delay for other errors
-                delay(1500)
-                try {
-                    NetworkBackendClient.sendJsonRequestWithCookies(
-                        url = url,
-                        method = "POST",
-                        headers = headers,
-                        options = RequestOptions(accessToken = token)
-                    )
-                } catch (e: Exception) {
-                    throw RuntimeException("Connection request failed with status code ${e.message}")
-                }
-            }
-        }
-        network.WireTestLogger.getLog("Backend").info("Response of send connection request is $response")
-    }
-
-    suspend fun updateUniqueUsername(user: ClientUser, newUniqueUsername: String) {
-        val userNameAlreadRegisteredError = 409
-        val tryAvoidDuplicates = newUniqueUsername == user.uniqueUsername
-        var currentUsername = newUniqueUsername
-        var attempts = 0
-
-        while (true) {
-            try {
-                updateSelfHandle(user, currentUsername)
-                user.uniqueUsername = currentUsername
-                return
-            } catch (e: HttpRequestException) {
-                if (tryAvoidDuplicates && e.returnCode == userNameAlreadRegisteredError && attempts < 5) {
-                    currentUsername = ClientUser.sanitizedRandomizedHandle(user.firstName)
-                } else {
-                    throw e
-                }
-            }
-            attempts++
-        }
-    }
-
-    fun getSelfDeletingMessagesSettings(teamMember: ClientUser): JSONObject {
-        val teamId = Uri.encode(getTeamId(teamMember))
-        val url = "i/teams/$teamId/features/selfDeletingMessages".composeCompleteUrl()
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put(AUTHORIZATION, "${basicAuth.getEncoded()}")
-        }
-
-        val response = NetworkBackendClient.sendJsonRequestWithCookies(
-            url = URL(url),
-            method = "GET",
-            headers = headers,
-            options = RequestOptions()
-        )
-
-        return JSONObject(response.body)
-    }
-
-    suspend fun switchServiceForTeam(
-        ownerOrAdminUser: ClientUser,
-        teamId: String,
-        providerId: String,
-        serviceId: String,
-        isEnabled: Boolean
-    ) {
-        val token = getAuthToken(ownerOrAdminUser)
-        val url = URI("teams/$teamId/services/whitelist".composeCompleteUrl()).toURL()
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put(AUTHORIZATION, "${token?.type} ${token?.value}")
-            put(accept, "*/*")
-        }
-
-        val requestBody = JSONObject().apply {
-            put("id", serviceId)
-            put("provider", providerId)
-            put("whitelisted", isEnabled)
-        }
-
-        NetworkBackendClient.sendJsonRequestWithCookies(
-            url = url,
-            method = "POST",
-            headers = headers,
-            body = requestBody.toString(),
-            options = RequestOptions(
-                accessToken = token,
-                expectedResponseCodes = NumberSequence.Array(
-                    intArrayOf(HttpURLConnection.HTTP_OK, HttpURLConnection.HTTP_NO_CONTENT)
-                )
-            )
-        )
-    }
-
-    suspend fun addServiceToConversation(asUser: ClientUser, serviceName: String, conversation: Conversation) {
-        val teamId = conversation.teamId
-            ?: throw IllegalStateException("Conversation '${conversation.name}' has no team id.")
-        val service = getWhitelistedService(asUser, teamId, serviceName)
-        val token = getAuthToken(asUser)
-        val url = URI("conversations/${conversation.id}/bots".composeCompleteUrl()).toURL()
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put(AUTHORIZATION, "${token?.type} ${token?.value}")
-        }
-
-        val requestBody = JSONObject().apply {
-            put("service", service.getString("id"))
-            put("provider", service.getString("provider"))
-        }
-
-        try {
-            NetworkBackendClient.sendJsonRequestWithCookies(
-                url = url,
-                method = "POST",
-                headers = headers,
-                body = requestBody.toString(),
-                options = RequestOptions(
-                    accessToken = token,
-                    expectedResponseCodes = NumberSequence.Array(
-                        intArrayOf(HttpURLConnection.HTTP_CREATED, HttpURLConnection.HTTP_NO_CONTENT)
-                    )
-                )
-            )
-        } catch (e: HttpRequestException) {
-            throw HttpRequestException(
-                "POST $url failed with HTTP ${e.returnCode}: ${e.message}",
-                e.returnCode
-            )
-        }
-    }
-
-    private suspend fun getWhitelistedService(asUser: ClientUser, teamId: String, serviceName: String): JSONObject {
-        val token = getAuthToken(asUser)
-        val url = URI(
-            "teams/$teamId/services/whitelisted?prefix=${Uri.encode(serviceName)}".composeCompleteUrl()
-        ).toURL()
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put(AUTHORIZATION, "${token?.type} ${token?.value}")
-        }
-
-        val response = NetworkBackendClient.sendJsonRequestWithCookies(
-            url = url,
-            method = "GET",
-            headers = headers,
-            options = RequestOptions(accessToken = token)
-        )
-
-        val services = JSONObject(response.body).getJSONArray("services")
-        for (index in 0 until services.length()) {
-            val service = services.getJSONObject(index)
-            if (service.optString("name") == serviceName) {
-                return service
-            }
-        }
-        throw NoSuchElementException("Service '$serviceName' is not whitelisted for team '$teamId'.")
-    }
-
-    fun getUserNameByID(domain: String, id: String, user: ClientUser): String {
-        val token = runBlocking { getAuthToken(user) }
-
-        val url = "users/$domain/$id/".composeCompleteUrl()
-        val headers = defaultheaders.toMutableMap().apply {
-            put(AUTHORIZATION, "${token?.type} ${token?.value}")
-        }
-
-        val response = NetworkBackendClient.sendJsonRequestWithCookies(
-            url = URL(url),
-            method = "GET",
-            headers = headers,
-            options = RequestOptions(accessToken = token)
-        )
-        return JSONObject(response.body).getString("name")
-    }
-
-    private suspend fun updateSelfHandle(user: ClientUser, handle: String) {
-        val token = getAuthToken(user)
-        val url = URL("self/handle".composeCompleteUrl())
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put("Authorization", "${token?.type} ${token?.value}")
-        }
-
-        val requestBody = JSONObject().apply {
-            put("handle", handle)
-        }
-
-        NetworkBackendClient.sendJsonRequestWithCookies(
-            url = url,
-            method = "PUT",
-            headers = headers,
-            body = requestBody.toString(),
-            options = RequestOptions(accessToken = token)
-        )
-    }
-
-    fun isDevelopmentApiEnabled(user: ClientUser): Boolean {
-        return getFeatureConfig("mls", user).get("status").equals("enabled")
-    }
-
-    suspend fun enableMLSFeatureTeam(
-        team: Team,
-        defaultCipherSuite: Int,
-        allowedCipherSuites: List<Int>,
-        defaultProtocol: String,
-        allowedProtocols: List<String>
-    ) {
-        val teamId = Uri.encode(team.id)
-        val url = URI("i/teams/$teamId/features/mls".composeCompleteUrl()).toURL()
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put("Authorization", basicAuth.getEncoded())
-        }
-
-        val requestBody = JSONObject().apply {
-            put("status", "enabled")
-            put(
-                "config",
-                JSONObject().apply {
-                    put("defaultCipherSuite", defaultCipherSuite)
-                    put("allowedCipherSuites", JSONArray(allowedCipherSuites))
-                    put("defaultProtocol", defaultProtocol)
-                    put("protocolToggleUsers", JSONArray())
-                    put("supportedProtocols", JSONArray(allowedProtocols))
-                }
-            )
-        }
-
-        NetworkBackendClient.sendJsonRequestWithCookies(
-            url = url,
-            method = "PUT",
-            headers = headers,
-            body = requestBody.toString(),
-            options = RequestOptions(
-                expectedResponseCodes = NumberSequence.Array(intArrayOf(HttpURLConnection.HTTP_OK))
-            )
-        )
-    }
-
-    suspend fun enableChannelFeatureViaBackdoorTeam(team: Team) {
-        val teamId = Uri.encode(team.id)
-        val headers = defaultheaders.toMutableMap().apply {
-            put("Authorization", basicAuth.getEncoded())
-        }
-
-        NetworkBackendClient.sendJsonRequestWithCookies(
-            url = URI("i/teams/$teamId/features/channels".composeCompleteUrl()).toURL(),
-            method = "PATCH",
-            headers = headers,
-            body = JSONObject().put("status", "enabled").toString(),
-            options = RequestOptions(
-                expectedResponseCodes = NumberSequence.Array(intArrayOf(HttpURLConnection.HTTP_OK))
-            )
-        )
-    }
-
-    suspend fun unlockChannelFeature(team: Team) {
-        val teamId = Uri.encode(team.id)
-        val url = URI("i/teams/$teamId/features/channels/unlocked".composeCompleteUrl()).toURL()
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put("Authorization", basicAuth.getEncoded())
-        }
-
-        NetworkBackendClient.sendJsonRequestWithCookies(
-            url = url,
-            method = "PUT",
-            headers = headers,
-            body = JSONObject().toString(),
-            options = RequestOptions(
-                expectedResponseCodes = NumberSequence.Array(intArrayOf(HttpURLConnection.HTTP_OK))
-            )
-        )
-    }
-
-    suspend fun unlockConferenceCallingFeature(team: Team) {
-        val teamId = Uri.encode(team.id)
-        val url = URI("i/teams/$teamId/features/conferenceCalling/unlocked".composeCompleteUrl()).toURL()
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put("Authorization", basicAuth.getEncoded())
-        }
-
-        NetworkBackendClient.sendJsonRequestWithCookies(
-            url = url,
-            method = "PUT",
-            headers = headers,
-            body = JSONObject().toString(),
-            options = RequestOptions(
-                expectedResponseCodes = NumberSequence.Array(intArrayOf(HttpURLConnection.HTTP_OK))
-            )
-        )
-    }
-
-    suspend fun enableConferenceCallingBackdoorViaBackendTeam(team: Team) {
-        val teamId = Uri.encode(team.id)
-        val url = URI("i/teams/$teamId/features/conferenceCalling".composeCompleteUrl()).toURL()
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put("Authorization", basicAuth.getEncoded())
-        }
-
-        val requestBody = JSONObject().apply {
-            put("status", "enabled")
-        }
-
-        NetworkBackendClient.sendJsonRequestWithCookies(
-            url = url,
-            method = "PATCH",
-            headers = headers,
-            body = requestBody.toString(),
-            options = RequestOptions(
-                expectedResponseCodes = NumberSequence.Array(intArrayOf(HttpURLConnection.HTTP_OK))
-            )
-        )
-    }
-
-    suspend fun disableConferenceCallingBackdoorViaBackendTeam(team: Team) {
-        val teamId = Uri.encode(team.id)
-        val url = URI("i/teams/$teamId/features/conferenceCalling".composeCompleteUrl()).toURL()
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put("Authorization", basicAuth.getEncoded())
-        }
-
-        val requestBody = JSONObject().apply {
-            put("status", "disabled")
-        }
-
-        NetworkBackendClient.sendJsonRequestWithCookies(
-            url = url,
-            method = "PUT",
-            headers = headers,
-            body = requestBody.toString(),
-            options = RequestOptions(
-                expectedResponseCodes = NumberSequence.Array(intArrayOf(HttpURLConnection.HTTP_OK))
-            )
-        )
-    }
-
-    suspend fun enableConferenceCallingViaBackendPersonalUser(personalUser: ClientUser) {
-        val userId = Uri.encode(personalUser.id)
-        val url = URI("i/users/$userId/features/conferenceCalling".composeCompleteUrl()).toURL()
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put("Authorization", basicAuth.getEncoded())
-        }
-
-        val requestBody = JSONObject().apply {
-            put("status", "enabled")
-        }
-
-        NetworkBackendClient.sendJsonRequestWithCookies(
-            url = url,
-            method = "PUT",
-            headers = headers,
-            body = requestBody.toString(),
-            options = RequestOptions(
-                expectedResponseCodes = NumberSequence.Array(intArrayOf(HttpURLConnection.HTTP_OK))
-            )
-        )
-    }
-
-    suspend fun upgradeToEnterprisePlanResult(team: Team) {
-        enableConferenceCallingBackdoorViaBackendTeam(team)
-    }
-
-    suspend fun getCallConfig(): JSONObject {
-        val url = URI("calls/config/v2".composeCompleteUrl()).toURL()
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put("Authorization", basicAuth.getEncoded())
-        }
-
-        val response = NetworkBackendClient.sendJsonRequestWithCookies(
-            url = url,
-            method = "GET",
-            headers = headers,
-            options = RequestOptions(
-                expectedResponseCodes = NumberSequence.Array(intArrayOf(HttpURLConnection.HTTP_OK))
-            )
-        )
-
-        return JSONObject(response.body)
-    }
-
-    suspend fun getPropertyValues(user: ClientUser): JSONObject {
-        val token = getAuthToken(user)
-        val url = URI("properties-values".composeCompleteUrl()).toURL()
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put("Authorization", "${token?.type} ${token?.value}")
-        }
-
-        val response = NetworkBackendClient.sendJsonRequestWithCookies(
-            url = url,
-            method = "GET",
-            headers = headers,
-            body = null,
-            options = RequestOptions(
-                accessToken = token,
-                expectedResponseCodes = NumberSequence.Array(intArrayOf(HttpURLConnection.HTTP_OK, HttpURLConnection.HTTP_NOT_FOUND))
-            )
-        )
-
-        return if (response.body.isNotEmpty()) {
-            JSONObject(response.body)
-        } else {
-            JSONObject()
-        }
-    }
-
     fun String.composeCompleteUrl(): String {
         return "${backendUrl}$this"
-    }
-
-    suspend fun acceptAllIncomingConnectionRequests(asUser: ClientUser) {
-        updateConnections(asUser, ConnectionStatus.Pending, ConnectionStatus.Accepted, null)
-    }
-
-    suspend fun acceptIncomingConnectionRequest(asUser: ClientUser, fromUser: ClientUser) {
-        updateConnections(
-            asUser,
-            ConnectionStatus.Pending,
-            ConnectionStatus.Accepted,
-            listOf(fromUser.id.orEmpty())
-        )
-    }
-
-    private suspend fun updateConnections(
-        asUser: ClientUser,
-        srcStatus: ConnectionStatus,
-        dstStatus: ConnectionStatus,
-        forUserIds: List<String>? = null
-    ) {
-        getAllConnections(asUser)
-            .filter { it.status == srcStatus && (forUserIds == null || forUserIds.contains(it.to)) }
-            .forEach { connection ->
-                try {
-                    changeConnectRequestStatus(asUser, connection.to.orEmpty(), connection.domain.orEmpty(), dstStatus)
-                } catch (e: Exception) {
-                    throw RuntimeException("Failed to update connection for ${connection.to}", e)
-                }
-            }
-    }
-
-    private suspend fun getAllConnections(user: ClientUser): List<Connection> {
-        var pagingState: String? = null
-        val result = mutableListOf<Connection>()
-
-        do {
-            val connectionsInfo = getConnectionsInfo(getAuthToken(user), pagingState)
-            val connections = connectionsInfo.getJSONArray("connections")
-
-            for (i in 0 until connections.length()) {
-                result.add(Connection.fromJSON(connections.getJSONObject(i)))
-            }
-
-            // Backward-compat with older backends
-            if (connectionsInfo.has("paging_state")) {
-                pagingState = connectionsInfo.getString("paging_state")
-            }
-        } while (connectionsInfo.getBoolean("has_more"))
-
-        return result
-    }
-
-    private fun getConnectionsInfo(token: AccessToken?, pagingState: String?): JSONObject {
-        val url = URL("list-connections".composeCompleteUrl())
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put(AUTHORIZATION, "${token?.type} ${token?.value}")
-        }
-
-        val requestBody = JSONObject().apply {
-            put("paging_state", pagingState)
-        }
-
-        return try {
-            val output = NetworkBackendClient.sendJsonRequestWithCookies(
-                url = url,
-                method = "POST",
-                headers = headers,
-                body = requestBody.toString(),
-                options = RequestOptions(accessToken = token)
-            )
-            JSONObject(output.body.toString())
-        } catch (e: Exception) {
-            e.printStackTrace()
-            if (e.message?.contains("404") == true) {
-                // Fallback for old backend
-                val fallbackUrl = URL("connections".composeCompleteUrl())
-                val output = NetworkBackendClient.sendJsonRequestWithCookies(
-                    url = fallbackUrl,
-                    method = "GET",
-                    headers = headers,
-                    options = RequestOptions(accessToken = token)
-                )
-                JSONObject(output.body.toString())
-            } else {
-                throw RuntimeException("Failed to fetch connections info", e)
-            }
-        }
-    }
-
-    suspend fun changeConnectRequestStatus(
-        asUser: ClientUser,
-        connectionId: String,
-        domain: String,
-        newStatus: ConnectionStatus
-    ) {
-        val token = getAuthToken(asUser)
-        val url = URL("connections/$domain/$connectionId".composeCompleteUrl())
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put(AUTHORIZATION, "${token?.type} ${token?.value}")
-        }
-
-        val requestBody = JSONObject().apply {
-            put("status", newStatus.toString())
-        }
-
-        try {
-            NetworkBackendClient.sendJsonRequestWithCookies(
-                url = url,
-                method = "PUT",
-                headers = headers,
-                body = requestBody.toString(),
-                options = RequestOptions(accessToken = token)
-            )
-        } catch (e: Exception) {
-            if (e.message?.contains("404") == true) {
-                // fallback for old backend
-                val fallbackUrl = URL("connections/$connectionId".composeCompleteUrl())
-                NetworkBackendClient.sendJsonRequestWithCookies(
-                    url = fallbackUrl,
-                    method = "PUT",
-                    headers = headers,
-                    body = requestBody.toString(),
-                    options = RequestOptions(accessToken = token)
-                )
-            } else {
-                throw RuntimeException("Failed to change connection status for $connectionId", e)
-            }
-        }
-    }
-
-    suspend fun enableForceAppLockFeature(team: Team, seconds: Int) {
-        val teamId = Uri.encode(team.id)
-        val url = URI("i/teams/$teamId/features/appLock".composeCompleteUrl()).toURL()
-
-        val headers = defaultheaders.toMutableMap().apply {
-            put("Authorization", basicAuth.getEncoded())
-        }
-
-        val requestBody = JSONObject().apply {
-            put("status", "enabled")
-            put(
-                "config",
-                JSONObject().apply {
-                    put("enforceAppLock", true)
-                    put("inactivityTimeoutSecs", seconds)
-                }
-            )
-        }
-
-        NetworkBackendClient.sendJsonRequestWithCookies(
-            url = url,
-            method = "PUT",
-            headers = headers,
-            body = requestBody.toString(),
-            options = RequestOptions(
-                expectedResponseCodes = NumberSequence.Array(intArrayOf(HttpURLConnection.HTTP_OK))
-            )
-        )
-    }
-
-    suspend fun disableConsentPopup(user: ClientUser) {
-        val privacyProperty = JSONObject().apply {
-            put("improve_wire", false)
-            put("marketing_consent", false)
-            put("telemetry_data_sharing", false)
-        }
-        val settingsProperty = JSONObject().apply {
-            put("privacy", privacyProperty)
-        }
-        val properties = JSONObject().apply {
-            put("settings", settingsProperty)
-        }
-        setPropertyValue(user, "webapp", properties.toString())
-    }
-
-    suspend fun setPropertyValue(user: ClientUser, propertyKey: String, properties: String) {
-        val url = "properties/$propertyKey".composeCompleteUrl()
-        val token = getAuthToken(user)
-        val headers = defaultheaders.toMutableMap().apply {
-            put("Authorization", "${token?.type} ${token?.value}")
-        }
-        NetworkBackendClient.sendJsonRequestWithCookies(
-            url = URL(url),
-            method = "PUT",
-            headers = headers,
-            body = properties,
-            options = RequestOptions(
-                expectedResponseCodes = NumberSequence.Array(intArrayOf(HttpURLConnection.HTTP_OK))
-            )
-        )
-    }
-}
-
-enum class ConnectionStatus {
-    Accepted, Blocked, Pending, Ignored, Sent, Cancelled;
-
-    override fun toString(): String = name.lowercase()
-
-    companion object {
-        fun fromString(s: String): ConnectionStatus =
-            entries.firstOrNull { it.name.equals(s, ignoreCase = true) }
-                ?: throw IllegalArgumentException("Connection status '$s' is unknown")
     }
 }
